+++ src/headline_scrapers/scrapers.py
@@ -1,5 +1,5 @@
 from headline_scrapers.base import BaseScraper
-from headline_scrapers.utils import clean_headlines
+from rumour_milled.utils import clean_headlines
 from storage.storage import HeadlineStorage
 from playwright.sync_api import TimeoutError
 import logging
